backend/requirements.txt
email-validator==2.1.1
fastapi==0.110.1
motor==3.3.1
passlib[bcrypt]==1.7.4
pydantic==2.6.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
requests==2.31.0
uvicorn[standard]==0.27.1
